requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
pytest>=7.4.3
httpx>=0.25.1
pytest-asyncio>=0.21.1
python-dotenv>=1.0.0
pgvector
langchain
langchain-community
langchain-ollama
python-dotenv
